# Enable HOMEBREW_BOOTSNAP for developers