--- a/manpages/brew.1
+++ b/manpages/brew.1
@@ -2398,7 +2398,7 @@ Use this username when accessing the Bintray API (where bottles are stored)\.
 \fBHOMEBREW_BOOTSNAP\fR
 .
 .br
-If set, use Bootsnap to speed up repeated \fBbrew\fR calls\. A no\-op when using Homebrew\'s vendored, relocatable Ruby on macOS (as it doesn\'t work)\.
+If set, use Bootsnap to speed up repeated \fBbrew\fR calls\. A no\-op in various configurations where it doesn\'t work\. Enabled by default if HOMEBREW_DEVELOPER is set or a developer command has been run\.
 .
 .TP
 \fBHOMEBREW_BOTTLE_DOMAIN\fR