- Enable it by default if you've run a developer command or set `HOMEBREW_DEVELOPER`.
- Clarify the documentation that there's various configuration in which it doesn't work.

I've rolled this out of most of GitHub's developers with no ill effects and we've enabled it in both CI and for various maintainers so it feels appropriate to have a bigger rollout. @Homebrew/brew: any thoughts?