@fxcoudert 2021-02-23
I've just experimented with it:

- it doesn't speed up the most common operation for mer, which is `brew up` (360 ms on my system, with or without `HOMEBREW_BOOTSNAP`, when I'm already up-to-date)
- `brew config` goes down from 0.8 s to 0.6 s
- `brew ls` goes down from 0.55 s to 0.35 s
- `brew cleanup` (with nothing to clean up) goes **up** from 1.0 s to 1.2 s