@MikeMcQuaid 2021-02-24
> * it doesn't speed up the most common operation for mer, which is `brew up` (360 ms on my system, with or without `HOMEBREW_BOOTSNAP`, when I'm already up-to-date)
> * `brew cleanup` (with nothing to clean up) goes **up** from 1.0 s to 1.2 s

These are IO bound and `brew cleanup` removes the existing bootsnap cache.

> * `brew config` goes down from 0.8 s to 0.6 s
> * `brew ls` goes down from 0.55 s to 0.35 s

The speedup here is mostly due to speeding up `require` pretty dramatically.