@Bo98 2021-02-24
Just as a FYI: https://github.com/Homebrew/discussions/discussions/923

(I think it might be fixable with some SDKROOT stuff, but I need more information on their setups first.)