@Bo98 2021-02-24
Actually maybe we should have a `DevelopmentTools.installed?` check here. I suspect maybe they are running without them (since ARM macOS updates delete the CLT).